pipeline/video-to-json/requirements.txt
# Video Processing
opencv-python>=4.8.0

# OCR (Optical Character Recognition)
pytesseract>=0.3.10

# Image Processing
Pillow>=10.0.0
imagehash>=4.3.1

# Numerical Operations
numpy>=1.24.0

# Text Similarity
scikit-learn>=1.3.0
python-Levenshtein>=0.21.1

# Progress Display
tqdm>=4.66.0
